requirements.txt
opencv-python==4.5.5.64
numpy==1.24.0
PyYAML==6.0
pure-python-adb==0.3.0.dev0
pywin32==305
schedule==1.1.0
pillow==9.5.0
pyautogui==0.9.53